% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bjack_cov}
\alias{bjack_cov}
\title{bjack_cov}
\arguments{
\item{stat1}{Vector of block-jackknife values for the first stat}

\item{stat2}{Vector of block-jackknife values for the second stat}
}
\value{
Covariance values
}
\description{
Compute the block-jackknife covariance between two stats
}
\details{
Compute the block-jackknife covariance between two stats with correction
}
\examples{
#
}
