% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{block_sum}
\alias{block_sum}
\alias{.block_sum}
\title{block_sum}
\usage{
.block_sum(stat, snp_bj_id)
}
\arguments{
\item{stat}{vector of n stat values}

\item{snp_bj_id}{integer n-length vector with block index (from 0 to nblock-1) of the stat value}
}
\value{
Return a vector of length nblocks containing the per-block sums of the input stat
}
\description{
Sugar to compute the sum of a stat per block
}
\details{
Sugar to compute the sum of a stat per block
}
\examples{
#
}
