% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{block_sum2}
\alias{block_sum2}
\alias{.block_sum2}
\title{block_sum2}
\usage{
.block_sum2(stat, snp_bj_id)
}
\arguments{
\item{stat}{vector of n stat values}

\item{snp_bj_id}{integer matrix of dim nblocks x 2 giving for each block the start and end stat value index}
}
\value{
Return a vector of length nblocks containing the per-block sums of the input stat
}
\description{
Sugar to compute the sum of a stat per block defined by a range of SNPs (allow treating overlapping blocks)
}
\details{
Sugar to compute the sum of a stat per block defined by a range of SNPs (allow treating overlapping blocks)
}
\examples{
#
}
