% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_Ddenom}
\alias{compute_Ddenom}
\alias{.compute_Ddenom}
\title{compute_Ddenom}
\usage{
.compute_Ddenom(snpQ2, f2idx, verbose)
}
\arguments{
\item{snpQ2}{the nsnp by (npop*(npop-1))/2 matrix of all pairwise Q2 estimates}

\item{f2idx}{a matrix of nDstat by 2 giving the index of the Q2 required to compute the denominator of the different F4}

\item{verbose}{if TRUE progression bar is printed on the terminal}
}
\value{
Return a vector of the denominator of the nDstat
}
\description{
Compute the denominator of Dstats
}
\details{
Compute the denominator of Dstats
}
\examples{
#
}
