% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_Ddenom_bjmeans}
\alias{compute_Ddenom_bjmeans}
\alias{.compute_Ddenom_bjmeans}
\title{compute_Ddenom_bjmeans}
\usage{
.compute_Ddenom_bjmeans(snpQ2, f2idx, snp_bj_id, verbose)
}
\arguments{
\item{snpQ2}{the nsnp by (npop*(npop-1))/2 matrix of all pairwise Q2 estimates}

\item{f2idx}{a matrix of nDstat by 2 giving the index of the Q2 required to compute the Dstat denominator}

\item{snp_bj_id}{integer vector of length nsnp giving the block index of each SNP}

\item{verbose}{if TRUE progression bar is printed on the terminal}
}
\value{
Return a vector with the block-jackknife mean estimates of the Dstat denominator
}
\description{
Compute the the block-jackknife mean of Dstat denominator
}
\details{
Compute the the block-jackknife mean of Dstat denominator
}
\examples{
#
}
