% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_F2_bjmeans}
\alias{compute_F2_bjmeans}
\alias{.compute_F2_bjmeans}
\title{compute_F2_bjmeans}
\usage{
.compute_F2_bjmeans(snpQ1, snpQ2, q1_idx, snp_bj_id, verbose)
}
\arguments{
\item{snpQ1}{the nsnp by npop matrix of Q1 estimates}

\item{snpQ2}{the nsnp by (npop*(npop-1))/2 matrix of all pairwise Q2 estimates}

\item{q1_idx}{the nsnp by 2 matrix with the indexes of the Q1 needed to compute each F2}

\item{snp_bj_id}{integer vector of length nsnp giving the block index of each SNP}

\item{verbose}{if TRUE progression bar is printed on the terminal}
}
\value{
Return a vector with the block-jackknife mean estimates of the F2 values
}
\description{
Compute the the block-jackknife mean of F2 values
}
\details{
Compute the the block-jackknife mean of F2 values
}
\examples{
#
}
