% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_F3fromF2}
\alias{compute_F3fromF2}
\alias{.compute_F3fromF2}
\title{compute_F3fromF2}
\usage{
.compute_F3fromF2(F2val, Hval, npops)
}
\arguments{
\item{F2val}{Numeric vector of length nF2=(npop*(npop-1))/2 with all pairwise F2 estimates}

\item{Hval}{Numeric vector of length npop with all within pop heterozygosity estimates}

\item{npops}{Integer giving the number of populations}
}
\value{
Return a matrix of length nF3=npops*(npops-1)*(npops-2)/2 rows and 2 columns corresponding to the F3 and F3star estimates
}
\description{
Compute all F3 from overall F2 values
}
\details{
Compute F3 and F3star estimates from F2 (and heterozygosities)
}
\examples{
#
}
