% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_F4fromF2}
\alias{compute_F4fromF2}
\alias{.compute_F4fromF2}
\title{compute_F4fromF2}
\usage{
.compute_F4fromF2(F2val, npops)
}
\arguments{
\item{F2val}{Numeric vector of length nF2=(npop*(npop-1))/2 with all pairwise F2 estimates}

\item{npops}{Integer giving the number of populations}
}
\value{
Return a vector of length nF4=(npops*(npops-1)/2) * ((npops-2)*(npops-3)/2) / 2 rows corresponding to all the F4 estimates for all possible configurations
}
\description{
Compute all F4 from overall F2 and Q2 values
}
\details{
Compute F4 from F2 (and heterozygosities)
}
\examples{
#
}
