% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_Q_bjmeans}
\alias{compute_Q_bjmeans}
\alias{.compute_Q_bjmeans}
\title{compute_Q_bjmeans}
\usage{
.compute_Q_bjmeans(snpQ, snp_bj_id, verbose)
}
\arguments{
\item{snpQ}{matrix of nsnp by nQ estimates of Q (e.g., Q1 or Q2)}

\item{snp_bj_id}{integer vector of length nsnp giving the block index of each SNP}

\item{verbose}{if TRUE progression bar is printed on the terminal}
}
\value{
Return a vector with the block-jackknife mean estimates for the nQ values
}
\description{
Compute the the block-jackknife mean of Q values
}
\details{
Compute the the block-jackknife mean of Q values
}
\examples{
#
}
