% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_snpHierFstAov}
\alias{compute_snpHierFstAov}
\alias{.compute_snpHierFstAov}
\title{compute_snpHierFstAov}
\usage{
.compute_snpHierFstAov(refcount, totcount, hapsize, popgrpidx, verbose)
}
\arguments{
\item{refcount}{Matrix of nsnpxnpop with counts (genotype or reads) for the reference allele}

\item{totcount}{Matrix of nsnpxnpop with total counts or read coverages}

\item{hapsize}{Vector of length npop giving the haploid size of each pool (if one element <=0, counts are interpreted as count data)}

\item{popgrpidx}{Vector of length npop giving the index (coded from 0 to ngrp-1) of the group of origin}

\item{verbose}{Logical (if TRUE progression bar is printed on the terminal)}
}
\value{
Return a nsnpsx6 matrix with SNP-specific MSI, MSP, MSG, nc, nc_p and nc_pp
}
\description{
Compute SNP-specific MSI, MSP, MSG, nc, nc_p and nc_pp used to derived the Anova estimator of hier. Fst for allele count or read count data (Pool-Seq)
}
\details{
Compute SNP-specific MSI, MSP, MSG, nc, nc_p and nc_pp used to derived the Anova estimator of hier. Fst for allele count or read count data (Pool-Seq)
}
\examples{
#
}
