% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_snpQ1}
\alias{compute_snpQ1}
\alias{.compute_snpQ1}
\title{compute_snpQ1}
\usage{
.compute_snpQ1(refcount, totcount, weight, verbose)
}
\arguments{
\item{refcount}{Matrix of nsnpxnpop with counts (genotype or reads) for the reference allele}

\item{totcount}{Matrix of nsnpxnpop with total counts or read coverages}

\item{weight}{Vector of length npop giving the weighting scheme (w=1 for allele count data and w=poolsize/(poolsize-1) for PoolSeq data)}

\item{verbose}{Logical (if TRUE progression bar is printed on the terminal)}
}
\value{
Return a vector of length nsnps with SNP-specific Q1
}
\description{
Compute SNP-specific Q1 by averaging over all samples
}
\details{
Compute all the SNP-specific Q1 over all pop. samples (useful for Fst computation with method Identity).
}
\examples{
#
}
