% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_snpQ1onepop}
\alias{compute_snpQ1onepop}
\alias{.compute_snpQ1onepop}
\title{compute_snpQ1onepop}
\usage{
.compute_snpQ1onepop(refcount, totcount, weight)
}
\arguments{
\item{refcount}{Vector of nsnp counts (genotype or reads) for the reference allele}

\item{totcount}{Vector of nsnp total counts or read coverages}

\item{weight}{Numeric (w=1 for allele count data and w=poolsize/(poolsize-1) for PoolSeq data)}
}
\value{
Return a vector of length nsnps with SNP-specific Q1
}
\description{
Compute SNP-specific Q1 for one pop
}
\details{
Compute SNP-specific Q1 for one pop. samples.
}
\examples{
#
}
