% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_snpQ2}
\alias{compute_snpQ2}
\alias{.compute_snpQ2}
\title{compute_snpQ2}
\usage{
.compute_snpQ2(refcount, totcount, pairs, verbose)
}
\arguments{
\item{refcount}{Matrix of nsnpxnpop with counts (genotype or reads) for the reference allele}

\item{totcount}{Matrix of nsnpxnpop with total counts or read coverages}

\item{pairs}{Matrix of npoppairsx2 giving the index for all the pairs of pops included in the computation}

\item{verbose}{Logical (if TRUE progression bar is printed on the terminal)}
}
\value{
Return a vector of length nsnps with SNP-specific Q2
}
\description{
Compute SNP-specific Q2 by averaging over all pairs of samples
}
\details{
Compute all the SNP-specific Q2 over all pop. pairs (useful for Fst computation with method Identity).
}
\examples{
#
}
