% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_snpQ2onepair}
\alias{compute_snpQ2onepair}
\alias{.compute_snpQ2onepair}
\title{compute_snpQ2onepair}
\usage{
.compute_snpQ2onepair(refcount1, refcount2, totcount1, totcount2)
}
\arguments{
\item{refcount1}{Vector of count (genotype or reads) for the reference allele in the first sample}

\item{refcount2}{Vector of count (genotype or reads) for the reference allele in the second sample}

\item{totcount1}{Vector of total count or read coverages in the first sample}

\item{totcount2}{Vector of total count or read coverages in the second sample}
}
\value{
Return a vector of length nsnps with SNP-specific Q1
}
\description{
Compute SNP-specific Q2 for a single pair of samples
}
\details{
Compute SNP-specific Q2 for a single pair of samples
}
\examples{
#
}
