% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_allele_names}
\alias{extract_allele_names}
\alias{.extract_allele_names}
\title{extract_allele_names}
\usage{
.extract_allele_names(allele_info, allele_idx)
}
\arguments{
\item{allele_info}{a character string vector (concatenated REF and ALT field of the vcf)}

\item{allele_idx}{Matrix with indexes of the two alleles of interest for the different markers}
}
\value{
Return a matrix with the two alleles after parsing the alleles info
}
\description{
Extract the alleles from the REF and ALT fields
}
\details{
Extract the alleles from the REF and ALT fields
}
\examples{
.extract_allele_names(c("A,C","A,C,T"),rbind(c(1,2),c(1,3)))

}
