% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.jacknife.blocks.R
\name{generate.jackknife.blocks}
\alias{generate.jackknife.blocks}
\title{Generate block coordinates for block-jackknife}
\usage{
generate.jackknife.blocks(x, nsnp.per.bjack.block, verbose = TRUE)
}
\arguments{
\item{x}{A pooldata or countdata object containing SNP positions (snp.info slot)}

\item{nsnp.per.bjack.block}{Number of consecutive SNPs of each block-jackknife block}

\item{verbose}{If TRUE extra information is printed on the terminal}
}
\value{
A list with the two following elements:
\enumerate{
\item "blocks.det": A matrix with three columns containing for each identified block (in row) the index of the start SNP, the index of the end SNP and the block Size in bp
\item "snp.block.id": A vector containing the blocks assigned to each SNP eligible for block-Jacknife (non eligible SNPs ares assigned NA)
\item "nblocks": A scalar corresponding to the number of blocks
\item "nsnps": Number of SNPs eligible for block-jackknife 'i.e., included in one block
}
}
\description{
Generate block coordinates for block-jackknife
}
