% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateF3names}
\alias{generateF3names}
\alias{.generateF3names}
\title{generateF3names}
\usage{
.generateF3names(popnames)
}
\arguments{
\item{popnames}{String vector with the names of all the pops}
}
\value{
Return a string matrix with 4 columns including the complete F3 configuration names (of the form Px;P1,P2), and the names of each pop involved in the configuration
}
\description{
Generate all names for F3 stats (same order as computation)
}
\details{
Generate all the npops*(npops-1)*(npops-2)/2 names for F3 stats (same order as computation)
}
\examples{
#
}
