% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateF4names}
\alias{generateF4names}
\alias{.generateF4names}
\title{generateF4names}
\usage{
.generateF4names(popnames)
}
\arguments{
\item{popnames}{String vector with the names of all the pops}
}
\value{
Return a string matrix with 5 columns including the complete F4 configuration names (of the form P1,P2;P3,P4), and the names of each pop involved in the configuration

#
}
\description{
Generate all names for F4 stats (same order as computation)
}
\details{
Generate all the nf4=(npops*(npops-1)/2)*((npops-2)*(npops-3)/2)/2 names for F4 stats (same order as computation)
}
