% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.graph.params.R
\name{is.graph.params}
\alias{is.graph.params}
\title{Check graph.params objects}
\usage{
is.graph.params(x)
}
\arguments{
\item{x}{The name (or a path) of the graph.params objet}
}
\description{
Check graph.params objects
}
