% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.objects.R
\name{is.pairwisefst}
\alias{is.pairwisefst}
\title{Check pairwisefst objects}
\usage{
is.pairwisefst(x)
}
\arguments{
\item{x}{The name (or a path) of the pairwisefst object}
}
\description{
Check pairwisefst objects
}
