% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.pooldata.R
\name{is.pooldata}
\alias{is.pooldata}
\title{Check pooldata objects}
\usage{
is.pooldata(x)
}
\arguments{
\item{x}{The name (or a path) of the Popoolation sync file (might be in compressed format)}
}
\description{
Check pooldata objects
}
