% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.example.files.R
\name{make.example.files}
\alias{make.example.files}
\title{Create example files}
\usage{
make.example.files(writing.dir = "")
}
\arguments{
\item{writing.dir}{Directory where to copy example files (e.g., set writing.dir=getwd() to copy in the current working directory)}
}
\description{
Write in the current directory example files corresponding to a sync (as obtained when parsing mpileup files with PoPoolation) and vcf (as obtained when parsing mpileup files with VarScan) gzipped files
}
\examples{
 make.example.files(writing.dir=tempdir())
}
