% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fitted.graph.R
\name{plot,fitted.graph-method}
\alias{plot,fitted.graph-method}
\title{plot pairwisefst object}
\usage{
\S4method{plot}{fitted.graph}(x, y)
}
\arguments{
\item{x}{Object of class fitted.graph}

\item{y}{dummy argument}
}
\description{
plot pairwisefst object
}
