% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.objects.R
\name{plot,fstats-method}
\alias{plot,fstats-method}
\title{plot fstats object}
\usage{
\S4method{plot}{fstats}(x, y, ...)
}
\arguments{
\item{x}{Object of class fstats}

\item{y}{dummy argument}

\item{...}{Other arguments to be passed to plot_fstats}
}
\description{
plot fstats object
}
\seealso{
see \code{\link{plot_fstats}} for details on plot_fstats arguments
}
