% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.graph.params.R
\name{plot,graph.params-method}
\alias{plot,graph.params-method}
\title{plot graph in graph.params object}
\usage{
\S4method{plot}{graph.params}(x, y)
}
\arguments{
\item{x}{Object of class fitted.graph}

\item{y}{dummy argument}
}
\description{
plot graph in graph.params object
}
