% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.objects.R
\name{plot,pairwisefst-method}
\alias{plot,pairwisefst-method}
\title{plot pairwisefst object}
\usage{
\S4method{plot}{pairwisefst}(x, y, ...)
}
\arguments{
\item{x}{Object of class pairwisefst}

\item{y}{dummy argument}

\item{...}{Some arguments to be passed to plot_fstats}
}
\description{
plot pairwisefst object
}
\seealso{
see \code{\link{plot_fstats}} for details on plot_fstats arguments
}
