% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poppair_idx}
\alias{poppair_idx}
\title{poppair_idx}
\arguments{
\item{idx_pop1}{Integer giving the (0-indexed) index of the first pop}

\item{idx_pop2}{Integer giving the (0-indexed) index of the second pop}

\item{nidx}{Integer giving the total number of indexes (i.e., number of pops)}
}
\value{
Return the (0-indexed) index for the row associated to the pairwise comparison in the ordered flat list of all (npop*(npop-1))/2 pairwise stats
}
\description{
Compute the index of the pairwise comparison from the idx of each pop
}
\details{
If idx_pop2 < idx_pop1, indexes are reversed
}
\examples{
#
}
