% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scan_allele_info}
\alias{scan_allele_info}
\alias{.scan_allele_info}
\title{scan_allele_info}
\usage{
.scan_allele_info(allele_info)
}
\arguments{
\item{allele_info}{a character string vector (ALT field of the vcf)}
}
\value{
Return a vector with two elements consisting i) the number of alleles (1+number of comma)
and ii) 0 or 1 if an indel is detected
}
\description{
Scan allele information in ALT field of a vcf
}
\details{
Scan allele information in ALT field of a vcf to identify the number of alleles and if there is indels
}
\examples{
.scan_allele_info(c("A,C","T","AAT"))

}
