% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.countdata.R
\name{show,countdata-method}
\alias{show,countdata-method}
\title{Show countdata object}
\usage{
\S4method{show}{countdata}(object)
}
\arguments{
\item{object}{Object of class countdata}
}
\description{
Show countdata object
}
