% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fitted.graph.R
\name{show,fitted.graph-method}
\alias{show,fitted.graph-method}
\title{Show fitted.graph object}
\usage{
\S4method{show}{fitted.graph}(object)
}
\arguments{
\item{object}{Object of class fitted.graph}
}
\description{
Show fitted.graph object
}
