% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.objects.R
\name{show,fstats-method}
\alias{show,fstats-method}
\title{Show fstats object}
\usage{
\S4method{show}{fstats}(object)
}
\arguments{
\item{object}{Object of class fstats}
}
\description{
Show fstats object
}
