% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.graph.params.R
\name{show,graph.params-method}
\alias{show,graph.params-method}
\title{Show graph.params object}
\usage{
\S4method{show}{graph.params}(object)
}
\arguments{
\item{object}{Object of class graph.params}
}
\description{
Show graph.params object
}
