% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.objects.R
\name{show,pairwisefst-method}
\alias{show,pairwisefst-method}
\title{Show pairwisefst object}
\usage{
\S4method{show}{pairwisefst}(object)
}
\arguments{
\item{object}{Object of class pairwisefst}
}
\description{
Show pairwisefst object
}
