% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.pooldata.R
\name{show,pooldata-method}
\alias{show,pooldata-method}
\title{Show pooldata object}
\usage{
\S4method{show}{pooldata}(object)
}
\arguments{
\item{object}{Object of class pooldata}
}
\description{
Show pooldata object
}
