% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolfstat-package.R
\docType{package}
\name{poolfstat}
\alias{poolfstat}
\alias{poolfstat-package}
\title{PoolFstat}
\description{
Functions for the computation of F-statistics from Pool-Seq data in population genomics studies. The package also includes several utilities to manipulate Pool-Seq data stored in standard format ('vcf' or 'rsync' files generated by the the 'PoPoolation' software) and perform conversion to alternative format (as used in the 'BayPass' and 'SelEstim' software).
}
\details{
Computing F-Statistics from Pool-Seq Data
}
