% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolfstat-package.R
\docType{package}
\name{poolfstat}
\alias{poolfstat}
\alias{poolfstat-package}
\title{PoolFstat}
\description{
Functions for the computation of F-, f- and D-statistics (e.g., Fst, hierarchical F-statistics, Patterson's F2, F3, F3*, F4 and D parameters) in population genomics studies from allele count or Pool-Seq read count data and for the fitting, building and visualization of admixture graphs. The package also includes several utilities to manipulate Pool-Seq data stored in standard format (e.g., such as 'vcf' files or 'rsync' files generated by the the 'PoPoolation' software) and perform conversion to alternative format (as used in the 'BayPass' and 'SelEstim' software). As of version 2.0, the package also includes utilities to manipulate standard allele count data (e.g., stored in TreeMix, BayPass and SelEstim format).
}
\details{
Computing f-Statistics and building admixture graphs based on allele count or Pool-Seq read count data
}
\author{
\strong{Maintainer}: Mathieu Gautier \email{mathieu.gautier@inrae.fr}

}
\keyword{internal}
